+++ dao/userDao.js
@@ -1,7 +1,8 @@
 'use strict';
 
 const Promise = require('bluebird');
-// var squel = require('squel').useFlavour('postgres');
+const Utils = require('../common/utils');
+var squel = require('squel').useFlavour('postgres');
 
 var userVo = null;
 var userScopeVo = null;
@@ -14,30 +15,69 @@ function _init (sequelize) {
   userAllowDeviceVo = sequelize.import('../models/userAllowDeviceVo');
   _sequelize = sequelize;
 }
-
-function _getUserList () {
-  return userVo.findAll();
+/**
+ * @param {Object} filterObj
+ * @param {Integer} filterObj.offset: starting index
+ * @param {Integer} filterObj.limit: data retrived
+ * @param {String} filterObj.userName: filter user name
+ * @param {String} filterObj.sortby: sort properties
+ * @param {String} filterObj.order: order asc or not
+ *  }
+ */
+function _getUserList (filterObj = {}) {
+  let sortby = filterObj.sortby ? filterObj.sortby : 'userId';
+  let order = typeof filterObj.order !== 'undefined' ? (filterObj.order ? 'ASC' : 'DESC') : 'ASC';
+  let filter = {
+    offset: filterObj.offset ? filterObj.offset : 0,
+    limit: filterObj.limit ? filterObj.limit : null,
+    order: _sequelize.literal(sortby + ' ' + order)
+  };
+  if (filterObj.userName) {
+    filter.where = {userName: filterObj.userName};
+  }
+  return userVo.findAndCountAll(filter);
 }
 
 function _getUserById (userId) {
-  return userVo.findOne({ where: {userId} });
+  let promises = [];
+  promises.push(userVo.findOne({ where: {userId} }));
+  promises.push(userScopeVo.findAll({ where: {userId} }));
+  return Promise.all(promises).then((result) => {
+    let response = {};
+    for (let res of result) {
+      if (Array.isArray(res)) {
+        response.scope = res.map((r) => r.scopeId);
+      } else {
+        response = Object.assign(response, res.dataValues);
+      }
+    }
+    return Promise.resolve(response);
+  });
 }
 
 function _getUserByName (userName) {
-  return userVo.findOne({where: {userName}});
+  return userVo.findOne({where:{userName}}).then((user) => {
+    if(!user){
+      return Promise.resolve(user);
+    }
+    user = user.dataValues;
+    return userScopeVo.findAll({where:{userId: user.userId}}).then((scopes) => {
+      user.scope = [].concat(scopes.map((s) => s.scopeId));
+      return Promise.resolve(user);
+    });
+  });
 }
 
 function _getUserScopeById (userId) {
   return userScopeVo.findAll({where: {userId}});
 }
 
-function _insertUser (userObj, trans) {
-  return userVo.findOne({where: {userName: userObj.userName}}).then((user) => {
-    if (user) {
-      Promise.reject(Error('userName is duplicated'));
-    }
-    return userVo.create(userObj, { transaction: trans });
-  });
+function _insertUser (users, trans) {
+  if (Array.isArray(users)) {
+    return userVo.bulkCreate(users, { transaction: trans });
+  } else {
+    return userVo.create(users, { transaction: trans });
+  }
 }
 
 function _insertUserScopeById (userId, scopeList, trans) {
@@ -58,7 +98,7 @@ function _updateUserScopeByName (userName, scopeList, trans) {
     for (let idx = 0; idx < scopeList.length; idx++) {
       array.push({userId: user.userId, scopeId: scopeList[idx]});
     }
-    return userScopeVo.destroy({where: {userId: user.userId}}, { transaction: trans }).then(function (c) {
+    return userScopeVo.destroy({ where: {userId: user.userId}, transaction: trans }).then(function (c) {
       return userScopeVo.bulkCreate(array, { transaction: trans });
     });
   });
@@ -69,13 +109,13 @@ function _updateUserScopeById (userId, scopeList, trans) {
   for (let idx = 0; idx < scopeList.length; idx++) {
     array.push({userId: userId, scopeId: scopeList[idx]});
   }
-  return userScopeVo.destroy({where: {userId}}, { transaction: trans }).then(function (c) {
+  return userScopeVo.destroy({ where: {userId}, transaction: trans }).then((c) => {
     return userScopeVo.bulkCreate(array, { transaction: trans });
   });
 }
 
 function _deleteUserById (userId, trans) {
-  return userVo.destroy({ where: { userId }, transaction: trans }).then(function (c) {
+  return userVo.destroy({ where: { userId }, transaction: trans }).then((c) => {
     return userScopeVo.destroy({ where: {userId}, transaction: trans }).then((result) => {
       return userAllowDeviceVo.destroy({ where: {userId}, transaction: trans });
     });